mt5_ai_trader_v2/requirements.txt
MetaTrader5>=5.0.4803
openai>=1.12.0
python-dotenv>=1.0.0
pandas>=2.1.4
numpy>=1.26.3
pandas-ta>=0.3.14b0
pydantic>=2.5.0
fastapi>=0.109.0
uvicorn>=0.27.0